Improved code samples

+++ 02/2_header_and_source_file/wallet.h
@@ -1,7 +1,7 @@
 #ifndef WALLET_H
 #define WALLET_H
 
-void deposit(unsigned int dollars, float cents);
-void withdraw(unsigned int dollars, float cents);
+void deposit(unsigned int dollars, unsigned int cents);
+void withdraw(unsigned int dollars, unsigned int cents);
 
 #endif /* WALLET_H */
